+++ src/pages/Bets/components/BetModal.tsx
@@ -1,14 +1,12 @@
-import { useContractWrite, useWaitForTransaction } from "wagmi";
 import ContractWriteResultCard from "../../../components/ContractWriteResultCard/ContractWriteResultCard_View";
 import Modal from "../../../components/Modal";
 import RequireWalletButton from "../../../components/RequireWalletButton/RequireWalletButton_View";
-import marketContractJson from "../../../abi/Market.json";
-import marketOracleContractJson from "../../../abi/MarketOracle.json";
 import { BetHistory } from "../../../types";
 import { Loader } from "../../../components";
 import useMarkets from "../../../hooks/market/useMarkets";
 import { useEffect, useState } from "react";
-import { EcSignature } from "../../../types/index";
+import useSettleContractWrite from "../../../hooks/market/useMarketSettle";
+import useMarketOracleResultWrite from "../../../hooks/market/useMarketOracle";
 
 type Props = {
   isOpen: boolean;
@@ -24,89 +22,6 @@ const BetModal = ({ isOpen, onClose, betData }: Props) => {
 };
 
 export default BetModal;
-
-type useSettleContractWriteArgs = {
-  marketAddress?: string;
-  index?: number;
-  raceResult?: boolean;
-  signature?: EcSignature;
-};
-const useSettleContractWrite = ({
-  marketAddress,
-  index
-}: useSettleContractWriteArgs) => {
-  const {
-    data,
-    error: settleBetError,
-    write: settleBetWrite
-  } = useContractWrite({
-    mode: "recklesslyUnprepared",
-    address: marketAddress || "",
-    abi: marketContractJson.abi,
-    functionName: "settle",
-    args: [index]
-  });
-
-  const settleBetTxHash = data?.hash;
-  const { isLoading: isSettleBetTxLoading, isSuccess: isSettleBetTxSuccess } =
-    useWaitForTransaction({
-      hash: settleBetTxHash
-    });
-  return {
-    settleBetWrite,
-    settleBetError,
-    isSettleBetTxLoading,
-    isSettleBetTxSuccess,
-    settleBetTxHash
-  };
-};
-
-type useMarketOracleResultWriteArgs = {
-  marketAddress?: string;
-  market_id?: string;
-  winningPropositionId?: string;
-  signature?: EcSignature;
-};
-const useMarketOracleResultWrite = ({
-  market_id,
-  winningPropositionId
-}: useMarketOracleResultWriteArgs) => {
-  const {
-    data,
-    error: marketOracleError,
-    write: setResultMarketOracleWrite
-  } = useContractWrite({
-    mode: "recklesslyUnprepared",
-    address:
-      process.env.REACT_APP_MARKET_ORACLE_CONTRACT ||
-      "0x592a44ebad029EBFff3Ee4950f1E74538a19a2ea",
-    abi: marketOracleContractJson.abi,
-    functionName: "setResult",
-    // TODO: Once we have switched the marketOracle contract to check EC signatures
-    // We can just pass the signature as the last argument
-    // For the moment we can just pass this blank signature to keep the contract happy
-    args: [
-      market_id,
-      winningPropositionId,
-      "0x0000000000000000000000000000000000000000000000000000000000000000"
-    ]
-  });
-
-  const marketOracleTxHash = data?.hash;
-  const {
-    isLoading: isMarketOracleTxHashTxLoading,
-    isSuccess: isMarketOracleTxHashTxSuccess
-  } = useWaitForTransaction({
-    hash: marketOracleTxHash
-  });
-  return {
-    setResultMarketOracleWrite,
-    marketOracleError,
-    isMarketOracleTxHashTxLoading,
-    isMarketOracleTxHashTxSuccess,
-    marketOracleTxHash
-  };
-};
 
 type SettlebetProps = {
   data?: BetHistory;
