+++ src/pages/Bets/components/BetTable.tsx
@@ -1,4 +1,5 @@
 import Skeleton from "react-loading-skeleton";
+import classnames from "classnames";
 import { BetHistory } from "../../../types";
 
 type Props = {
@@ -66,14 +67,28 @@ type RowProps = {
   onClick?: () => void;
 };
 const Row = ({ betData, onClick }: RowProps) => {
-  const rowStyle = () => {
-    if (betData?.settled) return "cursor-pointer bg-gray-300 hover:bg-gray-100";
-    if (betData?.winningPropositionId || betData?.marketResultAdded)
-      return "cursor-pointer bg-green-300 hover:bg-gray-100";
-    return "cursor-pointer hover:bg-gray-100";
-  };
+  // const rowStyle = () => {
+  //   if (betData?.settled) return "cursor-pointer bg-gray-300 hover:bg-gray-100";
+  //   if (betData?.winningPropositionId || betData?.marketResultAdded)
+  //     return "cursor-pointer bg-green-300 hover:bg-gray-100";
+  //   return "cursor-pointer hover:bg-gray-100";
+  // };
   return (
-    <tr key={betData?.proposition_id} onClick={onClick} className={rowStyle()}>
+    <tr
+      key={betData?.proposition_id}
+      onClick={onClick}
+      className={classnames(
+        "cursor-pointer hover:bg-gray-100",
+        {
+          "bg-green-300":
+            (betData?.winningPropositionId || betData?.marketResultAdded) &&
+            !betData?.settled
+        },
+        {
+          "bg-gray-300": betData?.settled
+        }
+      )}
+    >
       <td className="pl-5 pr-2 py-4 truncate">
         {betData?.punter ?? <Skeleton />}
       </td>
